+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
 * [mpm]: Improve table sorting with new version-aware tokenizer.
 * [gem]: Ignore ``default:`` prefix on package version parsing.
 * [mpm]: Remove ``packaging`` dependency. Rely on internal version parsing.
+* [mpm]: Allow stats to be printed for ``backup`` command.
 
 
 `3.0.0 (2020-03-25) <https://github.com/kdeldycke/meta-package-manager/compare/v2.9.0...v3.0.0>`_